Update s3download.py
We have to add something that runs the download_from_s3 here as well. I don't have time to figure out why the app.py was not working with the download_from_s3 call, so I just added this. Worst case it tries twice. 

Just have to patch the hole.